+++ mobile/test/performance/files_em_does_not_use
@@ -22,3 +22,4 @@ source/common/tls/ocsp/ocsp.h
 source/common/formatter/http_specific_formatter.h
 source/common/formatter/stream_info_formatter.h
 source/common/tls/default_tls_certificate_selector.h
+source/common/upstream/od_cds_api_impl.h
