+++ source/common/tcp_proxy/BUILD
@@ -88,6 +88,7 @@ envoy_cc_library(
         "//source/common/stream_info:stream_info_lib",
         "//source/common/stream_info:uint64_accessor_lib",
         "//source/common/upstream:load_balancer_context_base_lib",
+        "//source/common/upstream:od_cds_api_lib",
         "//source/extensions/upstreams/tcp/generic:config",
         "@envoy_api//envoy/config/accesslog/v3:pkg_cc_proto",
         "@envoy_api//envoy/extensions/filters/network/tcp_proxy/v3:pkg_cc_proto",
