+++ source/common/upstream/BUILD
@@ -85,7 +85,6 @@ envoy_cc_library(
         ":host_utility_lib",
         ":load_balancer_context_base_lib",
         ":load_stats_reporter_lib",
-        ":od_cds_api_lib",
         "//envoy/api:api_interface",
         "//envoy/config:xds_resources_delegate_interface",
         "//envoy/event:dispatcher_interface",
