+++ test/mocks/upstream/cluster_manager.cc
@@ -26,8 +26,8 @@ MockClusterManager::MockClusterManager()
   ON_CALL(*this, grpcAsyncClientManager()).WillByDefault(ReturnRef(async_client_manager_));
   ON_CALL(*this, localClusterName()).WillByDefault((ReturnRef(local_cluster_name_)));
   ON_CALL(*this, subscriptionFactory()).WillByDefault(ReturnRef(subscription_factory_));
-  ON_CALL(*this, allocateOdCdsApi(_, _, _))
-      .WillByDefault(Invoke([](const envoy::config::core::v3::ConfigSource&,
+  ON_CALL(*this, allocateOdCdsApi(_, _, _, _))
+      .WillByDefault(Invoke([](OdCdsCreationFunction, const envoy::config::core::v3::ConfigSource&,
                                OptRef<xds::core::v3::ResourceLocator>,
                                ProtobufMessage::ValidationVisitor&) -> OdCdsApiHandlePtr {
         return MockOdCdsApiHandle::create();
